% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnectMessage.R
\name{disconnectMessage}
\alias{disconnectMessage}
\title{Show a nice message when a shiny app disconnects}
\usage{
disconnectMessage(
  text = "An error occured. Please refresh the page and try again.",
  refresh = "Refresh",
  size = "m",
  background = "white",
  colour = "#444444",
  overlayColour = "black",
  overlayOpacity = 0.6
)
}
\arguments{
\item{text}{The text to show in the message.}

\item{refresh}{The text to show in a link that allows the user to refresh the page.
Use \code{refresh = ""} if you don't want to show a refresh link.}

\item{size}{The size of the message, one of "s", "m", "l" (small, medium, or large).}

\item{background}{The background colour of the message box.}

\item{colour}{The colour of the text of the message box.}

\item{overlayColour}{The colour of the overlay to draw on the page behind the message box.
An overlay is used to "grey out" the application and draw attention to the message. Use
\code{overlayOpacity = 0} to disable the overlay.}

\item{overlayOpacity}{The opacity of the overlay, from 0 (fully transparent) to 1
(fully opaque). Use \code{overlayOpacity = 0} to disable the overlay.}
}
\description{
A shiny app can disconnect for a variety of reasons: an unrecoverable error occurred in
the app, the server went down, the user's internet connection died, or any other reason
that might cause the shiny app to lose connection to its server. Call \code{disonnectMessage()}
anywhere in a Shiny app's UI to add a nice message when this happens. The message works both
locally (running Shiny apps within RStudio) and on Shiny servers (such as shinyapps.io,
RStudio Connect, Shiny Server Open Source, Shiny Server Pro).
}
\examples{
if (interactive()) {
  library(shiny)
  shinyApp(
    ui = fluidPage(
      disconnectMessage(),
      actionButton("disconnect", "Disconnect the app")
    ),
    server = function(input, output) {
      observeEvent(input$disconnect, {
        session$close()
      })
    }
  )
}
}
