% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnectMessage.R
\name{disconnectMessage2}
\alias{disconnectMessage2}
\title{Show a nice message when a shiny app disconnects or errors}
\usage{
disconnectMessage2()
}
\description{
This function is a version of \code{\link[=disconnectMessage]{disconnectMessage()}} with a pre-set combination
of parameters that results in a large centered message.
}
