% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydisconnect-package.R
\docType{package}
\name{shinydisconnect-package}
\alias{shinydisconnect}
\alias{shinydisconnect-package}
\title{shinydisconnect: Show a Nice Message When a 'Shiny' App Disconnects or Errors}
\description{
A 'Shiny' app can disconnect for a variety of reasons: an unrecoverable error occurred in the app, the server went down, the user lost internet connection, or any other reason that might cause the 'Shiny' app to lose connection to its server. With 'shinydisconnect', you can call disonnectMessage() anywhere in a Shiny app's UI to add a nice message when this happens. Works locally (running Shiny apps within 'RStudio') and on Shiny servers (such as shinyapps.io, 'RStudio Connect', 'Shiny Server Open Source', 'Shiny Server Pro'). See demo online at \url{https://daattali.com/shiny/shinydisconnect-demo/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daattali/shinydisconnect}
  \item \url{https://daattali.com/shiny/shinydisconnect-demo/}
  \item Report bugs at \url{https://github.com/daattali/shinydisconnect/issues}
}

}
\author{
\strong{Maintainer}: Dean Attali \email{daattali@gmail.com} (\href{https://orcid.org/0000-0002-5645-3493}{ORCID}) [copyright holder]

}
\keyword{internal}
